backend/requirements.txt
coverage==4.5.1
Django==1.9.9
djangorestframework==3.4.1
